combined contactViewModel and contactsViewModel into one VM
