+++ app/src/main/java/com/github/se/studybuddies/viewModels/ContactsViewModel.kt
@@ -1,8 +1,11 @@
 package com.github.se.studybuddies.viewModels
 
 import android.util.Log
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import com.github.se.studybuddies.data.Contact
 import com.github.se.studybuddies.data.ContactList
 import com.github.se.studybuddies.database.DatabaseConnection
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -14,6 +17,9 @@ class ContactsViewModel(private val uid: String? = null) : ViewModel() {
   private val _contacts = MutableStateFlow<ContactList>(ContactList(emptyList()))
   val contacts: StateFlow<ContactList> = _contacts
 
+  private val _contact = MutableLiveData(Contact.empty())
+  val contact: LiveData<Contact> = _contact
+
   init {
     if (uid != null) {
       fetchAllContacts(uid)
@@ -22,6 +28,17 @@ class ContactsViewModel(private val uid: String? = null) : ViewModel() {
 
   fun createContact(otherUID: String) {
     viewModelScope.launch { db.createContact(otherUID) }
+  }
+
+  fun fetchContactData(contactID: String) {
+    viewModelScope.launch { _contact.value = db.getContact(contactID) }
+  }
+
+  fun getOtherUser(contactID: String, uid: String): String {
+    fetchContactData(contactID)
+    return if ((contact.value?.members?.get(0) ?: "") == uid) {
+      contact.value?.members?.get(1) ?: ""
+    } else contact.value?.members?.get(0) ?: ""
   }
 
   fun fetchAllContacts(uid: String) {
